fixed a critical bug in cowsay and youtube_dl

+++ plugins/youtube_dl.lua
@@ -1,9 +1,14 @@
+-- Requires: (sudo) apt-get install youtube-dl
 function run(msg, matches)
   URL = matches[1]
   local receiver = get_receiver(msg)
+  if string.match(msg.text, '"') then
+	return 'Vergiss es'
+  else
     text = run_bash("rm tmp/video.mp4 && youtube-dl -o tmp/video.mp4 " .. URL)
 	send_video(get_receiver(msg), "tmp/video.mp4", ok_cb, false)
     return URL .. ' wurde gedownloadet und wird jetzt gesendet! \nDas kann je nach Video sehr lang dauern! \nAlso habt Gedult!'
+  end
 end
 
 return {
